fix: happy day scenario with hooks
