+++ src/app/modules/OpenshiftStreams/dialogs/CreateInstance/CreateInstanceWithSizes/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from './useAvailableProvidersAndDefault';
+export * from './useCreateInstance';
+export * from './useGetAvailableSizes';
